fix: address round-14 review — shared GPU discovery, prefix guard, doc alignment
Extract discoverGpuDevices() into test/integration/utils.js as a single
shared utility with process-wide caching.  Uses IndicTrans for device
probing (ggml device enumeration is device-dependent, not model-dependent)
so Bergamot and pivot-Bergamot tests get accurate GPU device lists instead
of always-empty results.

Fixes:
- R14-C1 BUG: bergamot discoverGpuDevices always returned empty (Bergamot
  is CPU-only, probe model must be IndicTrans)
- R14-C2 BUG: pivot-bergamot fr→es GPU discovery loaded wrong model pair
- R14-C3: shared discovery includes openclCacheDir setup for Android
- R14-A1 HIGH: eliminate 3-way copy-paste of discoverGpuDevices
- R14-A2/P2: single MAX_GPU_DEVICE_PROBES constant, single process cache
- R14-A4: console.warn → createLogger().warn
- R14-A5: align index.d.ts JSDoc (snake_case is primary, not camelCase)
- R14-S1: post-canonical /data/ prefix guard for backendsDir on Android
